Media Cloud Integration
File-Manager send a request to media cloud for each file uploaded.

From that request, file-manager gets the file preview  if there is one and create the base record of the media in the database.

File-manager has an endpoint that recieves the paths of generated files and save them in the according media entry.

If there is no preview returned, file-manager will asign a default one.

Some code was improved in the vue component javaScript